Issue 345 loki data (#358)
* temporary commit

* include for htons() needed on Linuxes, worked on Mac

* moved more stuff from LokiBase and into LokiInstrument

* found a overflow bug, corrected uint8_t to uint16_t, added trace

* PR review correction

+++ src/modules/loki/geometry/Config.h
@@ -31,7 +31,7 @@ public:
 
   std::vector<PanelGeometry> Panels;
   uint8_t NPanels{0}; // Loki panels, not logical geometry panels
-  uint8_t NTubesTotal{0}; // total number of tubes in instrument
+  uint16_t NTubesTotal{0}; // total number of tubes in instrument
   uint16_t Resolution{0}; /// Resolution along straws
   ESSGeometry * Geometry;
